Create the basic index.html file
I divided the webpage into five classes.
-.top
-.middle-1
-.middle-2
-.bottom
-.footer

For .top, I grouped the .header and the .hero-main-text together in .top so that I can display the two of them in the column direction. I also created .left-section and .right-section within the .hero-main-text so that one could be moved to the left and another to the right, respectively.

For .middle-1, I put a blank space between each pairs of <img>,<p> for clarity.

For .bottom, I again created a .left-section and .right-section to move one to the left and another to the right, respectively.

I'm using external css method for this project.