Add several declarations to .top
I set the height of the body to match the viewport (100vh). I used flex-direction: column so that I can push the footer to bottom of the page.

I used justify-content: space-between; and align-items: center; for .header, .hero so that I can align them to each other.

For some reason, .hero-main was automatically assigned a margin of 48px to top and bottom of the container. Set margin: 0; to get rid of existing margins.

I learned that the <button> element do not inherit their parent's font-family value by itself. I manually assign the font-family value of <button> to match that of the body.