Make changes to .bottom and its children
I removed "border: 10px" from .button under .top because it used the wrong syntax, it wasn't doing much anyways.

I also lower the contrast of .quote-text to make it so that it matches the picture more.

The rest of the additions are suited according to the given picture in the lesson. It should be self-explanatory ( I hope so).